backend/requirements.txt
fastapi
uvicorn
google-cloud-spanner
google-generativeai
firebase-admin
requests
python-dotenv
dspy-ai
